requirements.txt
alabaster==0.7.12
alembic==1.7.6
autodoc==0.5.0
Babel==2.9.1
beautifulsoup4==4.11.1
bidict==0.21.4
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
decorator==5.1.1
dnspython==2.2.1
docutils==0.17.1
Flask==2.0.3
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imagesize==1.3.0
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
pygame==2.0.3
Pygments==2.11.2
pyparsing==3.0.8
python-dotenv==0.20.0
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2022.1
requests==2.27.1
simple-websocket==0.5.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.32
urllib3==1.26.9
waitress==2.1.1
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.0.3
wsproto==1.1.0
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.4.0
